+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# Change Log
+# CodeSmile Extensions - Change Log
 
 #### v1.0.0 - Dec 26, 2023
 
-- First release
+- added Object.DestroyInAnyMode()
+- added Transform.DestroyInAnyMode() (destroys gameObject of transform)